+++ src/components/Datagrid/Datagrid.tsx
@@ -1,5 +1,6 @@
 import "@glideapps/glide-data-grid/dist/index.css";
 
+import useNavigator from "@dashboard/hooks/useNavigator";
 import { usePreventHistoryBack } from "@dashboard/hooks/usePreventHistoryBack";
 import DataEditor, {
   DataEditorProps,
@@ -86,7 +87,6 @@ export interface DatagridProps {
     defaultProps: Partial<ColumnPickerProps>,
   ) => ReactElement;
   onRowClick?: (item: Item) => void;
-  onRowHover?: (args: GridMouseEventArgs) => void;
   onColumnMoved?: (startIndex: number, endIndex: number) => void;
   onColumnResize?: (column: GridColumn, newSize: number) => void;
   readonly?: boolean;
@@ -95,6 +95,7 @@ export interface DatagridProps {
   freezeColumns?: DataEditorProps["freezeColumns"];
   verticalBorder?: DataEditorProps["verticalBorder"];
   columnSelect?: DataEditorProps["columnSelect"];
+  rowAnchor?: (item: Item) => string;
 }
 
 export const Datagrid: React.FC<DatagridProps> = ({
@@ -112,7 +113,6 @@ export const Datagrid: React.FC<DatagridProps> = ({
   onChange,
   renderColumnPicker,
   onRowClick,
-  onRowHover,
   getColumnTooltipContent,
   readonly = false,
   rowMarkers = "checkbox",
@@ -122,6 +122,7 @@ export const Datagrid: React.FC<DatagridProps> = ({
   onColumnMoved,
   onColumnResize,
   loading,
+  rowAnchor,
   hasRowHover = false,
   ...datagridProps
 }): ReactElement => {
@@ -130,6 +131,9 @@ export const Datagrid: React.FC<DatagridProps> = ({
   const datagridTheme = useDatagridTheme(readonly, readonly);
   const editor = useRef<DataEditorRef>();
   const customRenderers = useCustomCellRenderers();
+
+  const hackARef = useRef<HTMLAnchorElement>(null);
+  const navigate = useNavigator();
 
   const { scrolledToRight, scroller } = useScrollRight();
 
@@ -229,11 +233,24 @@ export const Datagrid: React.FC<DatagridProps> = ({
       if (hasRowHover) {
         setHoverRow(args.kind !== "cell" ? undefined : args.location[1]);
       }
-      if (onRowHover) {
-        onRowHover(args);
+
+      if (args.kind !== "cell" || !hackARef.current) {
+        return;
       }
+
+      const href = rowAnchor?.(args.location);
+
+      if (!href) {
+        return;
+      }
+
+      hackARef.current.style.left = `${window.scrollX + args.bounds.x}px`;
+      hackARef.current.style.width = `${args.bounds.width}px`;
+      hackARef.current.style.top = `${window.scrollY + args.bounds.y}px`;
+      hackARef.current.style.height = `${args.bounds.height}px`;
+      hackARef.current.href = href;
     },
-    [onRowHover, hasRowHover],
+    [hasRowHover, rowAnchor],
   );
 
   const handleGridSelectionChange = (gridSelection: GridSelection) => {
@@ -491,6 +508,16 @@ export const Datagrid: React.FC<DatagridProps> = ({
         clearTooltip={clearTooltip}
         bounds={tooltip?.bounds}
         title={tooltip?.title}
+      />
+      <a
+        ref={hackARef}
+        style={{ position: "absolute" }}
+        tabIndex={-1}
+        aria-hidden={true}
+        onClick={e => {
+          e.preventDefault();
+          navigate(e.currentTarget.pathname);
+        }}
       />
     </FullScreenContainer>
   );
