+++ src/products/components/ProductListPage/ProductListPage.tsx
@@ -287,9 +287,7 @@ export const ProductListPage: React.FC<ProductListPageProps> = props => {
             settings={settings}
             selectedChannelId={selectedChannelId}
             onUpdateListSettings={onUpdateListSettings}
-            onRowClick={id => {
-              navigate(productUrl(id));
-            }}
+            rowAnchor={productUrl}
           />
         ) : (
           <ProductListTiles
